Start rewrite of Hank to be better organized, supervised, and thought out
At the time of this commit, all Hank can do is connect to the server,
send the USER command, send the NICK command, and then IDENTIFY with
NickServ.

No parsing has been added yet. No message handling. All messages that
get sent are currently hardcoded. The next step is to implement a plugin
supervisor and plugins.